second attempt at integral calculation with window search
This algorithm is quite fun. We calculate the raw phase of E(x0)*h(x1-x0) and we try to move the window of the integral so that we only consider the portion where the phase has a small gradient. I still have to find a way for the window search to converge when no good point was found in the previous map. But at least now it works (sometimes...). Yay!